update catalog drawer design
