+++ locales/en/plugin__kubevirt-plugin.json
@@ -59,13 +59,14 @@
   "Create VirtualMachine from template": "Create VirtualMachine from template",
   "Name": "Name",
   "Virtual Machine Name": "Virtual Machine Name",
-  "Customize VirtualMachine": "Customize VirtualMachine",
-  "Quick create VirtualMachine": "Quick create VirtualMachine",
-  "This template supports quick create VirtualMachine": "This template supports quick create VirtualMachine",
   "VirtualMachine name": "VirtualMachine name",
   "Project": "Project",
   "Start this VirtualMachine after creation": "Start this VirtualMachine after creation",
+  "This template requires some additional parameters. Click the Customize Virtual machine button to complete the creation flow.": "This template requires some additional parameters. Click the Customize Virtual machine button to complete the creation flow.",
   "Quick create error": "Quick create error",
+  "Quick create VirtualMachine": "Quick create VirtualMachine",
+  "Customize VirtualMachine": "Customize VirtualMachine",
+  "This template supports quick create VirtualMachine": "This template supports quick create VirtualMachine",
   "Collapse": "Collapse",
   "Read more": "Read more",
   "N/A": "N/A",
